Refactor Database namespace
